Moved clearConsole() to a seperate file to improve readability of main code.
