+++ Bored/Bored.py
@@ -1,10 +1,11 @@
 # Console program that will return from a list; a random chore, a random fun thing to do, or a 
 # random selection from a list of both chores and fun things to do.
 #
-# V: 0.1.0
+# V: 0.1.1
 
 import sys
 import random
+from Helpers import * # Source of clearConsole()
 
 def menu():
 
@@ -53,27 +54,7 @@ def generateActivity() :
        
 def displayOutput(activity) :
     print(activity)
-
-
-def clearConsole(wait) : #function to clear console on Linux or Windows
-    """Clears console, with optional time delay.
-
-    Will attempt to clear the console for Windows, should that fail it will attempt to clear the
-    console for Linux.
-    """
-
-    import time
-    time.sleep(wait) 
-    # produces a delay based on the argument given to clearConsole()
-    
-    import os
-
-    try :
-       os.system('cls') #clears console on Windows
-
-    except :
-       os.system('clear') #clears console on Linux
-                
+     
 
 def main() :
