+++ Bored/Helpers.py
@@ -1,0 +1,19 @@
+def clearConsole(wait) : #function to clear console on Linux or Windows
+    """Clears console, with optional time delay.
+
+    Will attempt to clear the console for Windows, should that fail it will attempt to clear the
+    console for Linux.
+    """
+
+    import time
+    time.sleep(wait) 
+    # produces a delay based on the argument given to clearConsole()
+    
+    import os
+
+    try :
+       os.system('cls') #clears console on Windows
+
+    except :
+       os.system('clear') #clears console on Linux
+
